+++ app/(routes)/articles/[slug]/page.tsx
@@ -1,16 +1,7 @@
-import { Bebas_Neue } from "next/font/google";
-import cx from "classnames";
-
 import { client } from "@/sanity/client";
-import { FormattedDate } from "@components/FormattedDate";
-import { Tag } from "@components/Tag";
 import { BlockContentItemData } from "@customTypes/BlockContentTypes";
-import { BlockContent } from "@components/BlockContent";
 
-const bebasNeue = Bebas_Neue({
-  weight: "400",
-  subsets: ["latin"],
-});
+import { Article } from "@components/Article";
 
 const SLUGS_QUERY = `
 *[
@@ -49,7 +40,7 @@ type ArticleProps = {
     slug: string;
   };
 };
-export default async function Article({ params }: ArticleProps) {
+export default async function Page({ params }: ArticleProps) {
   const article = await client.fetch<ArticleData>(`
     *[
       _type == "blogEntry" &&
@@ -78,30 +69,12 @@ export default async function Article({ params }: ArticleProps) {
   } = article;
 
   return (
-    <div className="w-full flex flex-col items-center mt-12">
-      <div className="w-11/12 max-w-768 shadow-lg border-1 border-slate-600 p-8 mb-16">
-        <h1 className={cx("text-24 mb-8", bebasNeue.className)}>{title}</h1>
-
-        <h3 className="border-l-8 border-slate-800 pl-8 mb-12">
-          {description}
-        </h3>
-
-        <div className="flex flex-wrap mb-8">
-          {tags.map(({ value }, i) => (
-            <Tag label={value} key={value + i} />
-          ))}
-        </div>
-
-        <em className="w-full flex items-center text-12">
-          <hr className="flex-grow text-slate-500 mr-4" />
-          <FormattedDate value={createdAt} />
-        </em>
-      </div>
-
-      {/* Main content */}
-      <div className="w-11/12 max-w-768 shadow-lg border-1 border-slate-600 p-8 mb-16">
-        {<BlockContent blockContent={mainContent} />}
-      </div>
-    </div>
+    <Article
+      createdAt={createdAt}
+      title={title}
+      description={description}
+      tags={tags}
+      mainContent={mainContent}
+    />
   );
 }
